Create Cool Parallax Effect With CSS
You have probably seen the Parallax Effect used on websites with online ads. As you scroll down an article, the background image appears to scroll at a different pace, or an advert appears.

Alternatively, perhaps the background image changes as you visit different parts of the site. It's a cool effect that adds visual depth to the content and ideal even if you don't understand basic HTML code.