Create  Scrollable HTML Comment Box Code
This is a simple but helpful HTML element that lets you pack long snippets of text into a compact format. This way it doesn't take up the entire space on the page.

You can play with the colors and the size of the text box to make it fit your needs.